Made minor changes

+++ Jukebox/Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Diagnostics;
 using System.Threading;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 
 namespace Jukebox
 {
@@ -11,64 +13,64 @@ namespace Jukebox
             Song Mary = new Song("Mary had a little lamb")
             {
                 Sequence = new Note[] {
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.G4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.G4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Crotchet }
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.G4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.G4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Crotchet }
                 }
             };
 
             Song birthday = new Song("Happy birthday to you")
             {
                 Sequence = new Note[] {
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.A3, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.B3, Length = (int)NoteLength.Minim },
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.A3, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Minim },
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.G3, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.G4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.B3, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.A3, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.F4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.F4, Length = (int)NoteLength.Quaver },
-                new Note { Tone = SoundFrequency.E4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.D4, Length = (int)NoteLength.Crotchet },
-                new Note { Tone = SoundFrequency.C4, Length = (int)NoteLength.Minim },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.A3, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.B3, Length = NoteLength.Minim },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.A3, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Minim },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.G3, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.G4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.B3, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.A3, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.F4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.F4, Length = NoteLength.Quaver },
+                new Note { Tone = SoundFrequency.E4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.D4, Length = NoteLength.Crotchet },
+                new Note { Tone = SoundFrequency.C4, Length = NoteLength.Minim },
                 }
             };
 
@@ -82,7 +84,7 @@ namespace Jukebox
             foreach (var note in song.Sequence)
             {
                 Debug.WriteLine($"{note.Tone}: {note.Length}");
-                Console.Beep((int)note.Tone, 2000 / note.Length);
+                Console.Beep((int)note.Tone, 2000 / (int)note.Length);
             }
             Thread.Sleep(1000);
         }
